feat(model): Move package
